Parser was rewritten. One part of full dump implemented.
	Previous translating idea had failure. For new
idea parser had to be rewritten. Now it have to
walk through the tree (with help of recursive descending)
and save information about conditions for every action
(transition to another state, stopping any signal and another).
This information have to be saved and given to the next module
which would generate verilog code.
	One part of informative dump was implemented.